+++ svelte/demos/routes.js
@@ -10,6 +10,7 @@ import Buttons from "./cases/Buttons.svelte";
 import Values from "./cases/Values.svelte";
 import Ribbon from "./cases/Ribbon.svelte";
 import HeaderMenu from "./cases/HeaderMenu.svelte";
+import Tooltips from "./cases/Tooltips.svelte";
 
 export const links = [
 	["/base/:skin", "Toolbar basic", BasicInit, "BasicInit"],
@@ -22,6 +23,7 @@ export const links = [
 	],
 	["/ribbon/:skin", "Ribbon", Ribbon, "Ribbon"],
 	["/buttons/:skin", "Button types", Buttons, "Buttons"],
+	["/tooltip/:skin", "Button tooltips", Tooltips, "Tooltips"],
 	["/values/:skin", "Binding values", Values, "Values"],
 	["/menu/:skin", "Overflow menu", OverflowMenu, "OverflowMenu"],
 	[