+++ svelte/src/buttons/Icon.svelte
@@ -1,7 +1,8 @@
 <script>
 	import { Button } from "@svar-ui/svelte-core";
 
-	let { icon, text, css, type, disabled, menu, onclick } = $props();
+	let { icon, title, text, tooltip, css, type, disabled, menu, onclick } =
+		$props();
 </script>
 
 {#if menu}
@@ -12,7 +13,7 @@
 		{text}
 	</div>
 {:else}
-	<Button {icon} {type} {css} title={text} {disabled} {onclick} />
+	<Button {icon} {type} {css} {disabled} {title} {tooltip} {onclick} />
 {/if}
 
 <style>